+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 3.7.1 (unreleased)
 ------------------
 
-- Upgrade-step to reindex SearchableText for all contents with a table block.
+- Upgrade-step to reindex SearchableText for all contents with a table block (change in plone.restapi 8.16.1).
   [cekk]
 
 3.7.0 (2021-12-01)