+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-3.7.1 (unreleased)
+3.7.1 (2021-12-01)
 ------------------
 
 - Upgrade-step to reindex SearchableText for all contents with a table block (change in plone.restapi 8.16.1).